@@ -1,6 +1,8 @@
 
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { firstValueFrom } from 'rxjs';
+import { Covid } from '../models/covid.model';
 
 @Injectable({
   providedIn: 'root'
@@ -12,12 +14,13 @@ export class CovidapiService {
   constructor(private http: HttpClient) {
 
     this.apiUrl = `https://covid19-brazil-api.now.sh/api/report/v1/brazil/uf/pr`;
-   }
+  }
 
-    obterDados(): Promise<any>{
-      return this.http.get(`${this.apiUrl}`)
-      .toPromise()
-      .then( response => response);
-    }
+  obterDados(): Promise<Covid> {
+    return firstValueFrom(
+      this.http.get(`${this.apiUrl}`)
+    )
+      .then(response => response as Covid);
+  }
 
 }
